+++ .env.dist
@@ -1,3 +1,3 @@
 PORT=5000
 FLASK_ENV=development
-DATABASE_URL=sqlite:///sqitch/forum.db
+DATABASE_URL=sqlite:///../sqitch/forum.db
