+++ application/mod_users/models.py
@@ -1,14 +1,18 @@
 from sqlalchemy import Column, Integer, String, TIMESTAMP, func
-from .setup import db
+from ..main import db
 
-class User(db.Model):
-	__tablename__ = 'forum.user'
+class Base(db.Model):
+	__abstract__ = True
 
 	id = Column(Integer, primary_key=True, autoincrement=True)
+	created_at = Column(TIMESTAMP(timezone=True), server_default=func.now())
+	updated_at = Column(TIMESTAMP(timezone=True), server_default=func.now(), onupdate=func.now())
+
+class User(Base):
+	__tablename__ = 'forum.user'
+
 	username = Column(String(255), unique=True)
 	password_hash = Column(String(80))
-	created_at = Column(TIMESTAMP(timezone=True), server_default=func.now())
-	updated_at = Column(TIMESTAMP(timezone=True))
 
 	def __init__(self, username=None, password_hash=None):
 		self.username = username
