--- application/settings.py
@@ -1,2 +1,0 @@
-from dotenv import load_dotenv, find_dotenv
-load_dotenv(find_dotenv())
