+++ requirements.txt
@@ -3,7 +3,6 @@ bcrypt==3.1.7
 cffi==1.12.3
 Click==7.0
 Flask==1.1.1
-Flask-Cors==3.0.8
 Flask-SQLAlchemy==2.4.0
 isort==4.3.21
 itsdangerous==1.1.0
